Rename mail-helpers to support package
