+++ packages/client/src/helpers/response-error.js
@@ -3,49 +3,52 @@
 /**
  * Response error class
  */
-function ResponseError(response) {
+class ResponseError extends Error {
 
-  //Set response
-  this.response = response;
+  /**
+   * Constructor
+   */
+  constructor(response) {
 
-  //Set message
-  if (response.statusCode >= 500) {
-    this.message = 'Server error';
+    //Super
+    super();
+
+    //Set response
+    this.response = response;
+
+    //Set message
+    if (response.statusCode >= 500) {
+      this.message = 'Server error';
+    }
+    else if (response.statusCode >= 400) {
+      this.message = 'Client error';
+    }
+
+    //Capture stack trace
+    if (!this.stack) {
+      Error.captureStackTrace(this, this.constructor);
+    }
+
+    //Clean up stack trace
+    const regex = new RegExp(process.cwd() + '/', 'gi');
+    this.stack = this.stack.replace(regex, '');
   }
-  else if (response.statusCode >= 400) {
-    this.message = 'Client error';
+
+  /**
+   * Convert to string
+   */
+  toString() {
+    return `${this.name}: ${this.message}`;
   }
 
-  //Capture stack trace
-  if (!this.stack) {
-    Error.captureStackTrace(this, this.constructor);
+  /**
+   * Convert to simple object for JSON responses
+   */
+  toJSON() {
+    const {message, response} = this;
+    return {message, response};
   }
-
-  //Clean up stack trace
-  const regex = new RegExp(process.cwd() + '/', 'gi');
-  this.stack = this.stack.replace(regex, '');
 }
-
-/**
- * Extend prototype
- */
-ResponseError.prototype = Object.create(Error.prototype);
-ResponseError.prototype.constructor = ResponseError;
-
-/**
- * Convert to string
- */
-ResponseError.prototype.toString = function() {
-  return `${this.name}: ${this.message}`;
-};
-
-/**
- * Convert to simple object for JSON responses
- */
-ResponseError.prototype.toJSON = function() {
-  const {message, response} = this;
-  return {message, response};
-};
 
 //Export
 module.exports = ResponseError;
