+++ packages/inbound-mail-parser/src/parser.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs');
 const MailParser = require('mailparser').MailParser;
-const {Attachment} = require('@sendgrid/mail-helpers');
+const {Attachment} = require('@sendgrid/support');
 
 /**
  * Normalises attachment files retrieved from file system or parsed raw email
